Add subproject that contains the patchfiles as a regular library
Currently the patchfiles are copied into the source of each generated
parser, even though they are identical for all parsers and always use
a hardcoded package name. This means that all generated parsers
actually have conflicting classes, although in practice it's not
noticeable, because the classes are always identical.

This does however cause issues if the patchfiles are ever updated
(because then there could be multiple *different* copies of the same
classes on the classpath) and is incompatible with the Java 9+ module
system (which does not allow different modules to use the same package
name).

To fix these issues, the patchfiles will now be provided as a separate
runtime library, and in the future they will no longer be copied into
each generated parser.